third_party/blink/web_tests/platform/generic/fast/media/media-query-list-dom-events-expected.txt

Media query DOM events should fire (grouped by the MediaQueryList.)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS listenersCalled is expectedResult
PASS successfullyParsed is true

TEST COMPLETE
